+++ CHANGELOG.md
@@ -10,6 +10,7 @@ description: |-
 #### v0.7.4 (unreleased)
 - Add vkcs_networking_sdn data source for getting a list of available SDNs.
 - Add traffic_selector_ep_merge argument to vkcs_vpnaas_site_connection resource.
+- Add external_fixed_ips computed field to vkcs_networking_router resource and data source.
 
 #### v0.7.3
 - Add security_group_ids argument to vkcs_compute_instance resource